bug fixes on detail views

+++ app/src/main/java/me/ivg2/foodapp/FoodDetailFragment.java
@@ -81,7 +81,7 @@ public class FoodDetailFragment extends Fragment {
         index = arguments.getInt("index");
 
         etFoodName.setText(FoodItemRepository.get(index).getName());
-        etFoodQuantity.setText(Double.toString(FoodItemRepository.get(index).getQuantity()));
+        etFoodQuantity.setText(Integer.toString(FoodItemRepository.get(index).getQuantity()));
 
         //setting food expiration date to proper format
         DateTime targetDateTime = FoodItemRepository.get(index).getExpirationDate();
